chore: crates renaming, project structure
