--- app/build.rs
@@ -1,7 +1,0 @@
-/**
- * This is the build script for the app.
- * You don't need to change anything here.
- */
-fn main() {
-    automata_build_script::build_app();
-}
