+++ bin/zktls-pairs/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    automata_build_script::build_app();
+}
