+++ crates/enclave/src/error.rs
