+++ crates/untrusted-host/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mock-lib"
+name = "untrusted-host"
 version = "0.1.0"
 edition = "2021"
